--- src/routes/callback.ts
@@ -1,15 +1,0 @@
-/** @format */
-
-'use strict';
-
-import {Request, Response} from 'express';
-
-function callback(req: Request, res: Response) {
-    console.log('GOT CALLBACK');
-    console.log(req.body);
-    res.redirect("/")
-}
-
-export = {
-    // callback,
-};
