+++ src/services/encryption/open_id.ts
@@ -7,22 +7,23 @@ import optionService = require('../options');
 import myScryptService = require('./my_scrypt');
 import utils = require('../utils');
 import openIDEncryptionService = require('./open_id_encryption');
+import options = require('../options');
 
-function saveSubjectIdentifier(options: any, subjectIdentifier: string) {
+function saveSubjectIdentifier(subjectIdentifier: string) {
     if (isSubjectIdentifierSaved()) {
         throw new Error(
             "SubjectIdentifier is set already. Either change it or perform 'reset Subject Identifier' first."
         );
     }
-    options.setOption('subjectIdentifierVerificationSalt', utils.randomSecureToken(32), true);
-    options.setOption('subjectIdentifierDerivedKeySalt', utils.randomSecureToken(32), true);
+    options.setOption('subjectIdentifierVerificationSalt', utils.randomSecureToken(32));
+    options.setOption('subjectIdentifierDerivedKeySalt', utils.randomSecureToken(32));
     const subjectIdentifierVerificationKey = utils.toBase64(
         myScryptService.getSubjectIdentifierVerificationHash(subjectIdentifier)
     );
-    options.setOption('subjectIdentifierVerificationHash', subjectIdentifierVerificationKey, true);
+    options.setOption('subjectIdentifierVerificationHash', subjectIdentifierVerificationKey);
 
     openIDEncryptionService.setDataKey(subjectIdentifier, utils.randomSecureToken(16));
-    options.setOption('userSubjectIdentifierSaved', 'true', true);
+    options.setOption('userSubjectIdentifierSaved', 'true');
     return {
         success: true,
     };
