+++ src/views/login.ejs
@@ -48,9 +48,6 @@
                                         <button class="btn btn-success">Login</button>
                                     </div>
                     </form>
-                    <div>
-                        <button href="http://localhost:8080/authenticate">Login with OAuth</button>
-                    </div>
         </div>
     </div>
